+++ lib/src/api/devices.dart
@@ -1,4 +1,5 @@
 import 'dart:async';
+import 'dart:io' show Platform;
 
 import 'package:flutter/services.dart';
 
@@ -382,4 +383,84 @@ void onDeviceChange(OnDeviceChangeCallback? cb) {
 /// This must be called before any other function to work properly.
 Future<void> enableFakeMedia() async {
   await ffi.enableFakeMedia();
+}
+
+/// Foreground call service configuration.
+class ForegroundServiceConfig {
+  ForegroundServiceConfig({
+    this.enabled = true,
+    this.notificationOngoing = true,
+    this.notificationTitle = 'Ongoing call',
+    this.notificationText = 'Ongoing call',
+    this.notificationIcon = 'assets/icons/app_icon.png',
+  });
+
+  /// Indicator whether the foreground service is enabled.
+  bool enabled;
+
+  /// [ongoing] property of a notification.
+  ///
+  /// [ongoing]: https://tinyurl.com/ntfctn-doc#FLAG_ONGOING_EVENT
+  bool notificationOngoing;
+
+  /// [contentTitle] property of a notification.
+  ///
+  /// [contentTitle]: https://tinyurl.com/ntfctn-doc#EXTRA_TITLE
+  String notificationTitle;
+
+  /// [contentText] property of a notification.
+  ///
+  /// [contentText]: https://tinyurl.com/ntfctn-doc#EXTRA_TEXT
+  String notificationText;
+
+  /// [icon] property of a notification.
+  ///
+  /// This should be a full path to a bundled bitmap file. For the assets
+  /// configured in `pubspec.yaml` like this:
+  /// ```yaml
+  /// assets:
+  ///   - assets/icons/app_icon.png
+  /// ```
+  /// The full path would be `assets/icons/app_icon.png`.
+  ///
+  /// [ic_menu_call] will be used as a fallback if construction an icon from the
+  /// provided path fails.
+  ///
+  /// [icon]: https://tinyurl.com/ntfctn-doc#icon
+  /// [ic_menu_call]: https://tinyurl.com/andrawable#ic_menu_call
+  String notificationIcon;
+
+  /// Converts this model to the [Map] that can be transmitted via
+  /// [MethodChannel].
+  Map<String, dynamic> toMap() {
+    return {
+      'enabled': enabled,
+      'notificationOngoing': notificationOngoing,
+      'notificationTitle': notificationTitle,
+      'notificationText': notificationText,
+      'notificationIcon': notificationIcon,
+    };
+  }
+}
+
+/// Configures a foreground service and its notification on Android.
+///
+/// Foreground service is required for audio/video recording/playback to work
+/// when application is in the background.
+///
+/// Application will start foreground service whenever there is at least one
+/// active peer connection and stop when there are none anymore. This can also
+/// be controlled via [ForegroundServiceConfig.enabled] option. Foreground
+/// service can be stopped while running and restarted again by changing this
+/// parameter.
+///
+/// Foreground service notification parameters can be updated using this method
+/// at any moment. I.e., the title and the text could be changed while the
+/// notification is displayed and they would be updated immediately.
+Future<void> setupForegroundService(ForegroundServiceConfig config) async {
+  if (Platform.isAndroid) {
+    await _mediaDevicesMethodChannel.invokeMethod('setupForegroundService', {
+      'config': config.toMap(),
+    });
+  }
 }